+++ src/components/DataEntry/DataEntryTable/tests/index.test.tsx
@@ -51,7 +51,7 @@ jest.mock(
   "components/DataEntry/DataEntryTable/RecentEntry",
   () => MockRecentEntry
 );
-jest.mock("components/Pronunciations/PronunciationsComponent", () => "div");
+jest.mock("components/Pronunciations", () => "div");
 jest.mock("components/Pronunciations/Recorder");
 jest.mock("utilities/utilities");
