+++ src/components/Pronunciations/utilities.ts
@@ -1,5 +1,12 @@
 import { uploadAudio } from "backend";
-import { getFileNameForWord } from "components/Pronunciations/AudioRecorder";
+
+export function getFileNameForWord(wordId: string): string {
+  const fourCharParts = wordId.match(/.{1,6}/g);
+  const compressed = fourCharParts?.map((i) =>
+    Number("0x" + i).toString(36)
+  ) ?? ["unknownWord"];
+  return compressed.join("") + "_" + new Date().getTime().toString(36);
+}
 
 /*export async function uploadFile(wordId: string, file: File): Promise<string> {
   const fileName = getFileNameForWord(wordId);
