+++ src/goals/ReviewEntries/ReviewEntriesComponent/tests/CellColumns.test.tsx
@@ -12,7 +12,7 @@ import { newSemanticDomain } from "types/semanticDomain";
 import { newDefinition, newFlag, newGloss } from "types/word";
 import { Bcp47Code } from "types/writingSystem";
 
-jest.mock("components/Pronunciations/PronunciationsComponent", () => "div");
+jest.mock("components/Pronunciations", () => "div");
 jest.mock("i18next", () => {
   const i18n = jest.requireActual("i18next");
   return { ...i18n, t: (s: string) => s };
